requirements.txt
google-generativeai
python-dotenv
pydantic
faiss-cpu 
langchain
transformers
torch
pypdf
langchain-community